`get_prompt()` moved from `llm_connection` to `utils` and works from unit tests
