+++ pilot/prompts/prompts.py
@@ -2,8 +2,8 @@
 from utils.style import yellow
 from const import common
 from const.llm import MAX_QUESTIONS, END_RESPONSE
-from utils.llm_connection import create_gpt_chat_completion, get_prompt
-from utils.utils import capitalize_first_word_with_underscores, get_sys_message, find_role_from_step
+from utils.llm_connection import create_gpt_chat_completion
+from utils.utils import capitalize_first_word_with_underscores, get_sys_message, find_role_from_step, get_prompt
 from utils.questionary import styled_select, styled_text
 from logger.logger import logger
